refactor to put in correct utils.py
