+++ pyhealth/datasets/__init__.py
@@ -70,4 +70,10 @@ from .splitter import (
 )
 from .tuab import TUABDataset
 from .tuev import TUEVDataset
-from .utils import collate_fn_dict, collate_fn_dict_with_padding, get_dataloader
+from .utils import (
+    collate_fn_dict,
+    collate_fn_dict_with_padding,
+    get_dataloader,
+    load_processors,
+    save_processors,
+)
