+++ pyhealth/tasks/__init__.py
@@ -53,4 +53,3 @@ from .sleep_staging import (
 )
 from .sleep_staging_v2 import SleepStagingSleepEDF
 from .temple_university_EEG_tasks import EEG_events_fn, EEG_isAbnormal_fn
-from .utils import load_processors, save_processors
